Allow unconfined user filetransition for sudo log files
When "Defaults logfile=/var/log/sudo.log" is specified in sudo
configuration along with the documentation, an unconfined user needs
to have a named file transition defined to create the log file with a
proper type.

The sudo_filetrans_named_content_log() interface was created.

This is a follow-up on the 2ce62be20d63 ("Allow sudodomain use sudo.log
as a logfile") commit which was sufficient for confined administrators.

Resolves: rhbz#2164047